wx-rag-with-granite3/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
Authlib==1.3.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
grpcio==1.68.0
grpcio-health-checking==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.1.24
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.21
langchain-ibm==0.3.4
langchain-weaviate==0.0.3
langsmith==0.1.147
lomond==0.3.3
numpy==1.26.4
orjson==3.10.12
packaging==24.2
pandas==2.1.4
prompt_toolkit==3.0.48
protobuf==5.29.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.2
requests-toolbelt==1.0.0
setuptools==75.6.0
simsimd==4.4.0
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
wcwidth==0.2.13
weaviate-client==4.9.5
zipp==3.21.0
